# Faster generic dofs