Addresses a TODO in https://github.com/ONSAS/ONSAS.jl/pull/375

---

Before:

```julia
julia> @benchmark dofs($tetra)
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.254 μs … 285.858 μs  ┊ GC (min … max): 0.00% … 97.88%
 Time  (median):     1.312 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.487 μs ±   6.129 μs  ┊ GC (mean ± σ):  9.07% ±  2.19%

    ▆▄█▅
  ▂▆████▆▇▄▂▃▃▃▂▂▂▂▂▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  1.25 μs         Histogram: frequency by time        1.92 μs <

 Memory estimate: 2.41 KiB, allocs estimate: 26.
 ```

 After:

 ```julia
julia> @benchmark dofs($tetra)
BenchmarkTools.Trial: 10000 samples with 201 evaluations.
 Range (min … max):  395.935 ns … 19.777 μs  ┊ GC (min … max):  0.00% … 95.41%
 Time  (median):     582.299 ns              ┊ GC (median):     0.00%
 Time  (mean ± σ):   651.832 ns ±  1.344 μs  ┊ GC (mean ± σ):  16.05% ±  7.55%

  ▂                       █▇▂
  █▆▇▅▆▄▄▄▃▃▂▂▂▂▂▂▂▆▄▃▂▂▂▂███▆▅▄▃▂▂▄▄▃▃▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  396 ns          Histogram: frequency by time          850 ns <

 Memory estimate: 1.41 KiB, allocs estimate: 14.
```